requirements.txt
astor==0.8.1
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==3.0.0
cmake==3.25.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
einops==0.7.0
filelock==3.11.0
fonttools==4.44.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
json-tricks==3.17.3
kiwisolver==1.4.5
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.0
nni==3.0
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==9.3.0
prettytable==3.9.0
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
responses==0.24.0
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.3
simplejson==3.19.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torcheval==0.0.7
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typeguard==4.1.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
websockets==12.0
zipp==3.17.0
